requirements.txt
streamlit
torch
torchvision
joblib
Pillow
numpy
xgboost
scikit-learn